+++ src/pytorch_lightning/trainer/trainer.py
@@ -125,12 +125,8 @@ class Trainer:
         gradient_clip_val: Optional[Union[int, float]] = None,
         gradient_clip_algorithm: Optional[str] = None,
         num_nodes: int = 1,
-        num_processes: Optional[int] = None,  # TODO: Remove in 2.0
         devices: Optional[Union[List[int], str, int]] = None,
-        gpus: Optional[Union[List[int], str, int]] = None,  # TODO: Remove in 2.0
         auto_select_gpus: Optional[bool] = None,  # TODO: Remove in 2.0
-        tpu_cores: Optional[Union[List[int], str, int]] = None,  # TODO: Remove in 2.0
-        ipus: Optional[int] = None,  # TODO: Remove in 2.0
         enable_progress_bar: bool = True,
         overfit_batches: Union[int, float] = 0.0,
         track_grad_norm: Union[int, float, str] = -1,
@@ -235,19 +231,13 @@ class Trainer:
                 that don't support deterministic mode (requires PyTorch 1.11+). If not set, defaults to ``False``.
                 Default: ``None``.
 
-            devices: Will be mapped to either `gpus`, `tpu_cores`, `num_processes` or `ipus`,
-                based on the accelerator type.
+            devices: The devices to use. Can be set to a positive number (int or str), a sequence of device indices
+                (list or str), the value ``-1`` to indicate all available devices should be used, or ``"auto"`` for
+                automatic selection based on the chosen accelerator. Default: ``"auto"``.
 
             fast_dev_run: Runs n if set to ``n`` (int) else 1 if set to ``True`` batch(es)
                 of train, val and test to find any bugs (ie: a sort of unit test).
                 Default: ``False``.
-
-            gpus: Number of GPUs to train on (int) or which GPUs to train on (list or str) applied per node
-                Default: ``None``.
-
-                .. deprecated:: v1.7
-                    ``gpus`` has been deprecated in v1.7 and will be removed in v2.0.
-                    Please use ``accelerator='gpu'`` and ``devices=x`` instead.
 
             gradient_clip_val: The value at which to clip gradients. Passing ``gradient_clip_val=None`` disables
                 gradient clipping. If using Automatic Mixed Precision (AMP), the gradients will be unscaled before.
@@ -314,13 +304,6 @@ class Trainer:
             num_nodes: Number of GPU nodes for distributed training.
                 Default: ``1``.
 
-            num_processes: Number of processes for distributed training with ``accelerator="cpu"``.
-                Default: ``1``.
-
-                .. deprecated:: v1.7
-                    ``num_processes`` has been deprecated in v1.7 and will be removed in v2.0.
-                    Please use ``accelerator='cpu'`` and ``devices=x`` instead.
-
             num_sanity_val_steps: Sanity check runs n validation batches before starting the training routine.
                 Set it to `-1` to run all batches in all validation dataloaders.
                 Default: ``2``.
@@ -347,20 +330,6 @@ class Trainer:
 
             sync_batchnorm: Synchronize batch norm layers between process groups/whole world.
                 Default: ``False``.
-
-            tpu_cores: How many TPU cores to train on (1 or 8) / Single TPU to train on (1)
-                Default: ``None``.
-
-                .. deprecated:: v1.7
-                    ``tpu_cores`` has been deprecated in v1.7 and will be removed in v2.0.
-                    Please use ``accelerator='tpu'`` and ``devices=x`` instead.
-
-            ipus: How many IPUs to train on.
-                Default: ``None``.
-
-                .. deprecated:: v1.7
-                    ``ipus`` has been deprecated in v1.7 and will be removed in v2.0.
-                    Please use ``accelerator='ipu'`` and ``devices=x`` instead.
 
             track_grad_norm: -1 no tracking. Otherwise tracks that p-norm. May be set to 'inf' infinity-norm. If using
                 Automatic Mixed Precision (AMP), the gradients will be unscaled before logging them.
@@ -401,13 +370,9 @@ class Trainer:
         self._data_connector = DataConnector(self, multiple_trainloader_mode)
 
         self._accelerator_connector = AcceleratorConnector(
-            num_processes=num_processes,
             devices=devices,
-            tpu_cores=tpu_cores,
-            ipus=ipus,
             accelerator=accelerator,
             strategy=strategy,
-            gpus=gpus,
             num_nodes=num_nodes,
             sync_batchnorm=sync_batchnorm,
             benchmark=benchmark,