+++ src/pytorch_lightning/CHANGELOG.md
@@ -42,6 +42,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 - Removed `Trainer(strategy='horovod')` support ([#16150](https://github.com/Lightning-AI/lightning/pull/16150))
 
+- Removed legacy device arguments in Trainer ([#16171](https://github.com/Lightning-AI/lightning/pull/16171))
+  * Removed the `Trainer(gpus=...)` argument
+  * Removed the `Trainer(tpu_cores=...)` argument
+  * Removed the `Trainer(ipus=...)` argument
+  * Removed the `Trainer(num_processes=...)` argument
+
 
 ### Fixed